feat: rename `shared-types` to `types` in shared package
