+++ packages/shared/lib/utils/index.ts
@@ -1,2 +1,2 @@
 export * from './helpers.js';
-export type * from './shared-types.js';
+export type * from './types.js';
